Day 22 part 1 done
Started with the most logical thing - a matrix. Hated it, so went to
vectors of vectors. Better, but hated it. So switched to a sparse array,
and after battling with indices, found a awesome way to leverage
multiple dispatch to get named coordinates working with all the power of
sparse arrays, which have all the power of arrays.